Added e2e tests for custom domains (#4967)
* Added e2e tests for custom domains

* Added cleanup to newDiscussion test

* Added check to skip customDomain test if local
